chore: jsdoc and refactoring
